Addresses issue with validateReportingStructure (#74)
* Addresses issue with validateReportingStructure

- now clears messages from employeeId field when running function
- now returns all records, not just those with issues

Accounts for scenarios where data is updated and the function is re-run, messages are now accurate.

* Optimize validateReportingStructure function

Modified the validateReportingStructure function to return only modified records. This optimization reduces the amount of unnecessary data returned and streamlines the validation process. Removed redundant checks and streamlined error messages for better clarity.

validateReportingStructure to .ts

Bumped @flatfile/listener to 0.3.13

* Undo package lock changes

---------

Co-authored-by: John Mosesman <johnmosesman@gmail.com>